refactor: move deployment rejection tests to separate class
The CreateDeploymentTest became a bit large. I have moved the rejections to their own test class.
