+++ engine/src/test/java/io/camunda/zeebe/engine/processing/deployment/CreateDeploymentTest.java
@@ -8,7 +8,6 @@
 package io.camunda.zeebe.engine.processing.deployment;
 
 import static io.camunda.zeebe.protocol.Protocol.DEPLOYMENT_PARTITION;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 
@@ -18,10 +17,8 @@ import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
 import io.camunda.zeebe.model.bpmn.builder.ProcessBuilder;
 import io.camunda.zeebe.model.bpmn.instance.Message;
 import io.camunda.zeebe.protocol.record.Assertions;
-import io.camunda.zeebe.protocol.record.ExecuteCommandResponseDecoder;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.RecordType;
-import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.intent.DeploymentIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessIntent;
 import io.camunda.zeebe.protocol.record.value.DeploymentRecordValue;
@@ -31,9 +28,6 @@ import io.camunda.zeebe.test.util.Strings;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
 import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.Before;
@@ -44,7 +38,6 @@ import org.junit.Test;
 public final class CreateDeploymentTest {
 
   @ClassRule public static final EngineRule ENGINE = EngineRule.singlePartition();
-  private static final String TOO_LARGE_DEPLOYMENT_RESOURCE = "/processes/too_large_process.bpmn";
 
   @Rule
   public final RecordingExporterTestWatcher recordingExporterTestWatcher =
@@ -318,119 +311,6 @@ public final class CreateDeploymentTest {
   }
 
   @Test
-  public void shouldRejectDeploymentIfUsedInvalidMessage() {
-    // given
-    final BpmnModelInstance process =
-        Bpmn.createExecutableProcess().startEvent().intermediateCatchEvent("invalidMessage").done();
-
-    // when
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE.deployment().withXmlResource(process).expectRejection().deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment.getRecordType())
-        .isEqualTo(RecordType.COMMAND_REJECTION);
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfNotValidDesignTimeAspect() throws Exception {
-    // given
-    final Path path = Paths.get(getClass().getResource("/processes/invalid_process.bpmn").toURI());
-    final byte[] resource = Files.readAllBytes(path);
-
-    // when
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE.deployment().withXmlResource(resource).expectRejection().deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment)
-        .hasKey(ExecuteCommandResponseDecoder.keyNullValue())
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasIntent(DeploymentIntent.CREATE)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT);
-    assertThat(rejectedDeployment.getRejectionReason())
-        .contains("ERROR: Must have at least one start event");
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfNotValidRuntimeAspect() throws Exception {
-    // given
-    final Path path =
-        Paths.get(getClass().getResource("/processes/invalid_process_condition.bpmn").toURI());
-    final byte[] resource = Files.readAllBytes(path);
-
-    // when
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE.deployment().withXmlResource(resource).expectRejection().deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment)
-        .hasKey(ExecuteCommandResponseDecoder.keyNullValue())
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasIntent(DeploymentIntent.CREATE)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT);
-    assertThat(rejectedDeployment.getRejectionReason())
-        .contains("Element: flow2 > conditionExpression")
-        .contains("ERROR: failed to parse expression");
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfOneResourceIsNotValid() throws Exception {
-    // given
-    final Path path1 = Paths.get(getClass().getResource("/processes/invalid_process.bpmn").toURI());
-    final Path path2 = Paths.get(getClass().getResource("/processes/collaboration.bpmn").toURI());
-    final byte[] resource1 = Files.readAllBytes(path1);
-    final byte[] resource2 = Files.readAllBytes(path2);
-
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(resource1)
-            .withXmlResource(resource2)
-            .expectRejection()
-            .deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment)
-        .hasKey(ExecuteCommandResponseDecoder.keyNullValue())
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasIntent(DeploymentIntent.CREATE)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT);
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfNoResources() {
-    // when
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE.deployment().expectRejection().deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment)
-        .hasKey(ExecuteCommandResponseDecoder.keyNullValue())
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasIntent(DeploymentIntent.CREATE)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT);
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfNotParsable() {
-    // when
-    final Record<DeploymentRecordValue> rejectedDeployment =
-        ENGINE
-            .deployment()
-            .withXmlResource("not a process".getBytes(UTF_8))
-            .expectRejection()
-            .deploy();
-
-    // then
-    Assertions.assertThat(rejectedDeployment)
-        .hasKey(ExecuteCommandResponseDecoder.keyNullValue())
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasIntent(DeploymentIntent.CREATE)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT);
-  }
-
-  @Test
   public void shouldIncrementProcessVersions() {
     // given
     final BpmnModelInstance modelInstance =
@@ -585,81 +465,6 @@ public final class CreateDeploymentTest {
 
     assertDifferentResources(
         findProcess(originalProcesses, processId), findProcess(repeatedProcesses, processId));
-  }
-
-  @Test
-  public void shouldRejectDeploymentWithDuplicateResources() {
-    // given
-    final BpmnModelInstance definition1 =
-        Bpmn.createExecutableProcess("process1").startEvent().done();
-    final BpmnModelInstance definition2 =
-        Bpmn.createExecutableProcess("process2").startEvent().done();
-    final BpmnModelInstance definition3 =
-        Bpmn.createExecutableProcess("process2")
-            .startEvent()
-            .serviceTask("task", (t) -> t.zeebeJobType("j").zeebeTaskHeader("k", "v"))
-            .done();
-
-    // when
-    final Record<DeploymentRecordValue> deploymentRejection =
-        ENGINE
-            .deployment()
-            .withXmlResource("p1.bpmn", definition1)
-            .withXmlResource("p2.bpmn", definition2)
-            .withXmlResource("p3.bpmn", definition3)
-            .expectRejection()
-            .deploy();
-
-    // then
-    Assertions.assertThat(deploymentRejection)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT)
-        .hasRejectionReason(
-            "Expected to deploy new resources, but encountered the following errors:\n"
-                + "Duplicated process id in resources 'p2.bpmn' and 'p3.bpmn'");
-  }
-
-  @Test
-  public void shouldRejectDeploymentWithInvalidTimerStartEventExpression() {
-    // given
-    final BpmnModelInstance definition =
-        Bpmn.createExecutableProcess("process1")
-            .startEvent("start-event-1")
-            .timerWithCycleExpression("INVALID_CYCLE_EXPRESSION")
-            .done();
-
-    // when
-    final Record<DeploymentRecordValue> deploymentRejection =
-        ENGINE.deployment().withXmlResource("p1.bpmn", definition).expectRejection().deploy();
-
-    // then
-    Assertions.assertThat(deploymentRejection)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT)
-        .hasRejectionReason(
-            "Expected to deploy new resources, but encountered the following errors:\n"
-                + "'p1.bpmn': - Element: start-event-1\n"
-                + "    - ERROR: Invalid timer cycle expression ("
-                + "failed to evaluate expression "
-                + "'INVALID_CYCLE_EXPRESSION': no variable found for name "
-                + "'INVALID_CYCLE_EXPRESSION')\n");
-  }
-
-  @Test
-  public void shouldRejectDeploymentIfResourceIsTooLarge() {
-    // when
-    final Record<DeploymentRecordValue> deploymentRejection =
-        ENGINE
-            .deployment()
-            .withXmlClasspathResource(TOO_LARGE_DEPLOYMENT_RESOURCE)
-            .expectRejection()
-            .deploy();
-
-    // then
-    Assertions.assertThat(deploymentRejection)
-        .hasRejectionType(RejectionType.INVALID_ARGUMENT)
-        .hasRejectionReason(
-            "Unable to deploy resources as the size exceeds the maximum batch size. Please split "
-                + "the resources into separate deployments, or reduce the size of the deployed "
-                + "resource.");
   }
 
   private ProcessMetadataValue findProcess(
